+++ packages/flow-core/CHANGELOG.md
@@ -4,18 +4,32 @@
 
 # Change Log
 
+## [1.17.3] - 2023-07-03
+### Improvements
+
+- `f-select`: `optionTemplate` function has now a second value `isSelected`, to check and differ the template displayed on selection area.
+
+### Bug Fixes
+
+- `f-select`: on create a new option, the creation must be always auto selected; z-index of dropdown is fixed.
+- `f-progress-bar`: variant curved issue fixed.
+
 ## [1.17.2] - 2023-07-03
 ### Bug Fixes
 - tooltip not displaying when element is disabled.
 - `f-input` browser autofill style updated.
 - `f-text` state switching issue.
 
+
 ## [1.17.1] - 2023-06-21
+
 ### Improvements
+
 - Extract shared configuration into `@cldcvr/flow-core-config` package.
 - Allow loading of icon bundles asynchronously
 
 ### Bug Fixes
+
 - Fix theme subscription cleanup in f-icon
 
 ## [1.17.0] - 2023-06-14
@@ -23,6 +37,7 @@
 ### Features
 
 - `f-text-area` : `mask-value` property added to behave like password field.
+
 ## [1.16.2] - 2023-06-09
 
 ### Improvements